rag/llamaindex-rag/pyproject.toml
[project]
name = "llamaindex-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.15",
    "docx2txt>=0.9",
    "langchain-deepseek>=0.1.3",
    "llama-index>=0.12.50",
    "llama-index-core>=0.12.50",
    "llama-index-embeddings-ollama>=0.6.0",
    "llama-index-llms-langchain>=0.6.1",
    "llama-index-llms-ollama>=0.6.2",
    "llama-index-readers-file>=0.4.11",
    "llama-index-vector-stores-chroma>=0.4.2",
    "openai>=1.97.0",
    "python-dotenv>=1.1.1",
]
